pyproject.toml
[project]
name = "djangoba-tools"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "obayemi", email = "obayemi@ikjpz.com" }]
requires-python = ">=3.12"
dependencies = ["django>=5.1.3", "djangorestframework>=3.15.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
